feat: implement connected agents feature in user profile
- Added a new interface for connected agents and state management to fetch and display connected agents in the user profile.
- Implemented API call to retrieve connected agents based on the authenticated user, enhancing user experience by showing active agents.
- Introduced a profile requirements checklist for claiming agents, ensuring users meet necessary criteria before claiming.
- Updated Firestore API to include profile status checks, improving validation for agent claims based on user profile completeness and visibility.
