+++ components/members/MemberCard.tsx
@@ -1,0 +1,258 @@
+import Image from "next/image";
+import type { PublicMember } from "@/types/members";
+import { getInitials } from "@/lib/utils";
+
+interface MemberCardProps {
+  member: PublicMember;
+}
+
+export function MemberCard({ member }: MemberCardProps) {
+  const v = member.visibility;
+  const isAgent = member.memberType === "agent";
+
+  return (
+    <div className="bg-neutral-900 rounded-xl p-6 border border-neutral-800 hover:border-neutral-700 transition-colors">
+      {/* Header */}
+      <div className="flex items-start gap-4 mb-4">
+        <div className="relative shrink-0">
+          {member.photoURL ? (
+            <Image
+              src={member.photoURL}
+              alt={member.displayName || "Member"}
+              width={56}
+              height={56}
+              className="rounded-full object-cover"
+            />
+          ) : (
+            <div className={`w-14 h-14 rounded-full flex items-center justify-center text-white font-semibold text-lg ${
+              isAgent ? "bg-purple-900/50" : "bg-neutral-800"
+            }`}>
+              {isAgent ? (
+                <svg width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" aria-hidden="true">
+                  <rect x="3" y="11" width="18" height="10" rx="2" />
+                  <circle cx="12" cy="5" r="2" />
+                  <path d="M12 7v4" />
+                  <circle cx="8" cy="16" r="1" fill="currentColor" />
+                  <circle cx="16" cy="16" r="1" fill="currentColor" />
+                </svg>
+              ) : (
+                getInitials(member.displayName)
+              )}
+            </div>
+          )}
+        </div>
+        <div className="flex-1 min-w-0">
+          <div className="flex items-center gap-2 mb-0.5">
+            <h3 className="text-white font-semibold text-lg truncate">
+              {member.displayName || "Anonymous"}
+            </h3>
+            {/* Member Type Tag */}
+            <span className={`shrink-0 px-2 py-0.5 text-xs rounded-full font-medium ${
+              isAgent 
+                ? "bg-purple-500/10 text-purple-400 border border-purple-500/30" 
+                : "bg-emerald-500/10 text-emerald-400 border border-emerald-500/30"
+            }`}>
+              {isAgent ? "Agent" : "Human"}
+            </span>
+          </div>
+          {!isAgent && v?.showJobTitle && member.jobTitle && (
+            <p className="text-neutral-400 text-sm truncate">{member.jobTitle}</p>
+          )}
+          {!isAgent && v?.showCompany && member.company && (
+            <p className="text-neutral-500 text-sm truncate">{member.company}</p>
+          )}
+          {isAgent && member.owner?.displayName && v?.showOwner && (
+            <p className="text-neutral-500 text-sm truncate">
+              Owned by {member.owner.displayName}
+            </p>
+          )}
+        </div>
+      </div>
+
+      {/* Bio */}
+      {v?.showBio && member.bio && (
+        <p className="text-neutral-300 text-sm mb-4 line-clamp-3">{member.bio}</p>
+      )}
+
+      {/* Location */}
+      {v?.showLocation && member.location && (
+        <div className="flex items-center gap-2 text-neutral-400 text-sm mb-4">
+          <svg
+            xmlns="http://www.w3.org/2000/svg"
+            width="14"
+            height="14"
+            viewBox="0 0 24 24"
+            fill="none"
+            stroke="currentColor"
+            strokeWidth="2"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+            aria-hidden="true"
+          >
+            <path d="M20 10c0 6-8 12-8 12s-8-6-8-12a8 8 0 0 1 16 0Z" />
+            <circle cx="12" cy="10" r="3" />
+          </svg>
+          {member.location}
+        </div>
+      )}
+
+      {/* Badges */}
+      <div className="flex flex-wrap gap-2 mb-4">
+        {v?.showDiscord && member.discord && (
+          <span className="px-2 py-1 bg-[#5865F2]/10 text-[#5865F2] text-xs rounded-full inline-flex items-center gap-1">
+            <svg width="12" height="12" viewBox="0 0 24 24" fill="currentColor" aria-hidden="true">
+              <path d="M20.317 4.37a19.791 19.791 0 0 0-4.885-1.515.074.074 0 0 0-.079.037c-.21.375-.444.864-.608 1.25a18.27 18.27 0 0 0-5.487 0 12.64 12.64 0 0 0-.617-1.25.077.077 0 0 0-.079-.037A19.736 19.736 0 0 0 3.677 4.37a.07.07 0 0 0-.032.027C.533 9.046-.32 13.58.099 18.057a.082.082 0 0 0 .031.057 19.9 19.9 0 0 0 5.993 3.03.078.078 0 0 0 .084-.028 14.09 14.09 0 0 0 1.226-1.994.076.076 0 0 0-.041-.106 13.107 13.107 0 0 1-1.872-.892.077.077 0 0 1-.008-.128 10.2 10.2 0 0 0 .372-.292.074.074 0 0 1 .077-.01c3.928 1.793 8.18 1.793 12.062 0a.074.074 0 0 1 .078.01c.12.098.246.198.373.292a.077.077 0 0 1-.006.127 12.299 12.299 0 0 1-1.873.892.077.077 0 0 0-.041.107c.36.698.772 1.362 1.225 1.993a.076.076 0 0 0 .084.028 19.839 19.839 0 0 0 6.002-3.03.077.077 0 0 0 .032-.054c.5-5.177-.838-9.674-3.549-13.66a.061.061 0 0 0-.031-.03zM8.02 15.33c-1.183 0-2.157-1.085-2.157-2.419 0-1.333.956-2.419 2.157-2.419 1.21 0 2.176 1.096 2.157 2.42 0 1.333-.956 2.418-2.157 2.418zm7.975 0c-1.183 0-2.157-1.085-2.157-2.419 0-1.333.955-2.419 2.157-2.419 1.21 0 2.176 1.096 2.157 2.42 0 1.333-.946 2.418-2.157 2.418z"/>
+            </svg>
+            Discord
+          </span>
+        )}
+        {v?.showGithubBadge && member.github && (
+          <span className="px-2 py-1 bg-neutral-800/50 text-white text-xs rounded-full inline-flex items-center gap-1">
+            <svg width="12" height="12" viewBox="0 0 24 24" fill="currentColor" aria-hidden="true">
+              <path
+                fillRule="evenodd"
+                d="M12 2C6.477 2 2 6.484 2 12.017c0 4.425 2.865 8.18 6.839 9.504.5.092.682-.217.682-.483 0-.237-.008-.868-.013-1.703-2.782.605-3.369-1.343-3.369-1.343-.454-1.158-1.11-1.466-1.11-1.466-.908-.62.069-.608.069-.608 1.003.07 1.531 1.032 1.531 1.032.892 1.53 2.341 1.088 2.91.832.092-.647.35-1.088.636-1.338-2.22-.253-4.555-1.113-4.555-4.951 0-1.093.39-1.988 1.029-2.688-.103-.253-.446-1.272.098-2.65 0 0 .84-.27 2.75 1.026A9.564 9.564 0 0112 6.844c.85.004 1.705.115 2.504.337 1.909-1.296 2.747-1.027 2.747-1.027.546 1.379.202 2.398.1 2.651.64.7 1.028 1.595 1.028 2.688 0 3.848-2.339 4.695-4.566 4.943.359.309.678.92.678 1.855 0 1.338-.012 2.419-.012 2.747 0 .268.18.58.688.482A10.019 10.019 0 0022 12.017C22 6.484 17.522 2 12 2z"
+                clipRule="evenodd"
+              />
+            </svg>
+            GitHub
+          </span>
+        )}
+        {v?.showEventsAttended && member.eventsAttended && member.eventsAttended > 0 && (
+          <span className="px-2 py-1 bg-emerald-500/10 text-emerald-400 text-xs rounded-full">
+            {member.eventsAttended} event{member.eventsAttended !== 1 ? "s" : ""} attended
+          </span>
+        )}
+        {v?.showTalksGiven && member.talksGiven && member.talksGiven > 0 && (
+          <span className="px-2 py-1 bg-purple-500/10 text-purple-400 text-xs rounded-full">
+            {member.talksGiven} talk{member.talksGiven !== 1 ? "s" : ""} given
+          </span>
+        )}
+        {member.pullRequestsCount && member.pullRequestsCount > 0 && (
+          <span className="px-2 py-1 bg-blue-500/10 text-blue-400 text-xs rounded-full">
+            {member.pullRequestsCount} PR{member.pullRequestsCount !== 1 ? "s" : ""}
+          </span>
+        )}
+      </div>
+
+      {/* Social Links */}
+      <div className="flex items-center gap-1 pt-4 border-t border-neutral-800 -ml-2">
+        {v?.showWebsite && member.socialLinks?.website && (
+          <a
+            href={member.socialLinks.website}
+            target="_blank"
+            rel="noopener noreferrer"
+            aria-label="Website (opens in new tab)"
+            className="text-neutral-400 hover:text-white transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-white rounded p-2 min-w-[44px] min-h-[44px] flex items-center justify-center"
+          >
+            <svg
+              xmlns="http://www.w3.org/2000/svg"
+              width="18"
+              height="18"
+              viewBox="0 0 24 24"
+              fill="none"
+              stroke="currentColor"
+              strokeWidth="2"
+              strokeLinecap="round"
+              strokeLinejoin="round"
+              aria-hidden="true"
+            >
+              <circle cx="12" cy="12" r="10" />
+              <path d="M12 2a14.5 14.5 0 0 0 0 20 14.5 14.5 0 0 0 0-20" />
+              <path d="M2 12h20" />
+            </svg>
+          </a>
+        )}
+        {v?.showLinkedIn && member.socialLinks?.linkedIn && (
+          <a
+            href={member.socialLinks.linkedIn}
+            target="_blank"
+            rel="noopener noreferrer"
+            aria-label="LinkedIn (opens in new tab)"
+            className="text-neutral-400 hover:text-white transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-white rounded p-2 min-w-[44px] min-h-[44px] flex items-center justify-center"
+          >
+            <svg
+              xmlns="http://www.w3.org/2000/svg"
+              width="18"
+              height="18"
+              viewBox="0 0 24 24"
+              fill="currentColor"
+              aria-hidden="true"
+            >
+              <path d="M20.447 20.452h-3.554v-5.569c0-1.328-.027-3.037-1.852-3.037-1.853 0-2.136 1.445-2.136 2.939v5.667H9.351V9h3.414v1.561h.046c.477-.9 1.637-1.85 3.37-1.85 3.601 0 4.267 2.37 4.267 5.455v6.286zM5.337 7.433c-1.144 0-2.063-.926-2.063-2.065 0-1.138.92-2.063 2.063-2.063 1.14 0 2.064.925 2.064 2.063 0 1.139-.925 2.065-2.064 2.065zm1.782 13.019H3.555V9h3.564v11.452zM22.225 0H1.771C.792 0 0 .774 0 1.729v20.542C0 23.227.792 24 1.771 24h20.451C23.2 24 24 23.227 24 22.271V1.729C24 .774 23.2 0 22.222 0h.003z"/>
+            </svg>
+          </a>
+        )}
+        {v?.showTwitter && member.socialLinks?.twitter && (
+          <a
+            href={member.socialLinks.twitter}
+            target="_blank"
+            rel="noopener noreferrer"
+            aria-label="X/Twitter (opens in new tab)"
+            className="text-neutral-400 hover:text-white transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-white rounded p-2 min-w-[44px] min-h-[44px] flex items-center justify-center"
+          >
+            <svg
+              xmlns="http://www.w3.org/2000/svg"
+              width="18"
+              height="18"
+              viewBox="0 0 24 24"
+              fill="currentColor"
+              aria-hidden="true"
+            >
+              <path d="M18.244 2.25h3.308l-7.227 8.26 8.502 11.24H16.17l-5.214-6.817L4.99 21.75H1.68l7.73-8.835L1.254 2.25H8.08l4.713 6.231zm-1.161 17.52h1.833L7.084 4.126H5.117z"/>
+            </svg>
+          </a>
+        )}
+        {v?.showGithub && member.socialLinks?.github && (
+          <a
+            href={member.socialLinks.github}
+            target="_blank"
+            rel="noopener noreferrer"
+            aria-label="GitHub (opens in new tab)"
+            className="text-neutral-400 hover:text-white transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-white rounded p-2 min-w-[44px] min-h-[44px] flex items-center justify-center"
+          >
+            <svg
+              xmlns="http://www.w3.org/2000/svg"
+              width="18"
+              height="18"
+              viewBox="0 0 24 24"
+              fill="currentColor"
+              aria-hidden="true"
+            >
+              <path d="M12 0c-6.626 0-12 5.373-12 12 0 5.302 3.438 9.8 8.207 11.387.599.111.793-.261.793-.577v-2.234c-3.338.726-4.033-1.416-4.033-1.416-.546-1.387-1.333-1.756-1.333-1.756-1.089-.745.083-.729.083-.729 1.205.084 1.839 1.237 1.839 1.237 1.07 1.834 2.807 1.304 3.492.997.107-.775.418-1.305.762-1.604-2.665-.305-5.467-1.334-5.467-5.931 0-1.311.469-2.381 1.236-3.221-.124-.303-.535-1.524.117-3.176 0 0 1.008-.322 3.301 1.23.957-.266 1.983-.399 3.003-.404 1.02.005 2.047.138 3.006.404 2.291-1.552 3.297-1.23 3.297-1.23.653 1.653.242 2.874.118 3.176.77.84 1.235 1.911 1.235 3.221 0 4.609-2.807 5.624-5.479 5.921.43.372.823 1.102.823 2.222v3.293c0 .319.192.694.801.576 4.765-1.589 8.199-6.086 8.199-11.386 0-6.627-5.373-12-12-12z"/>
+            </svg>
+          </a>
+        )}
+        {v?.showSubstack && member.socialLinks?.substack && (
+          <a
+            href={member.socialLinks.substack}
+            target="_blank"
+            rel="noopener noreferrer"
+            aria-label="Substack (opens in new tab)"
+            className="text-neutral-400 hover:text-white transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-white rounded p-2 min-w-[44px] min-h-[44px] flex items-center justify-center"
+          >
+            <svg
+              xmlns="http://www.w3.org/2000/svg"
+              width="18"
+              height="18"
+              viewBox="0 0 24 24"
+              fill="currentColor"
+              aria-hidden="true"
+            >
+              <path d="M22.539 8.242H1.46V5.406h21.08v2.836zM1.46 10.812V24L12 18.11 22.54 24V10.812H1.46zM22.54 0H1.46v2.836h21.08V0z"/>
+            </svg>
+          </a>
+        )}
+        {v?.showMemberSince && member.createdAt && (
+          <span className="text-neutral-500 text-xs ml-auto">
+            Member since{" "}
+            {member.createdAt.toDate().toLocaleDateString("en-US", {
+              month: "short",
+              year: "numeric",
+            })}
+          </span>
+        )}
+      </div>
+    </div>
+  );
+}
