+++ components/members/MemberDirectory.tsx
@@ -1,0 +1,324 @@
+"use client";
+
+import { useState } from "react";
+import Link from "next/link";
+import { useRouter } from "next/navigation";
+import { useMembers } from "@/hooks/useMembers";
+import { MemberCard } from "./MemberCard";
+import { FilterCheckbox } from "./FilterCheckbox";
+import type { SortOption } from "@/types/members";
+
+interface MemberDirectoryProps {
+  initialSearch?: string;
+}
+
+export function MemberDirectory({ initialSearch = "" }: MemberDirectoryProps) {
+  const router = useRouter();
+  const {
+    members,
+    loading,
+    searchQuery,
+    setSearchQuery,
+    filters,
+    setFilters,
+    sortBy,
+    setSortBy,
+    filteredAndSortedMembers,
+    activeFilterCount,
+    clearFilters,
+  } = useMembers(initialSearch);
+
+  const [showFilters, setShowFilters] = useState(false);
+
+  const handleClearFilters = () => {
+    clearFilters();
+    router.push("/members", { scroll: false });
+  };
+
+  return (
+    <section className="py-8 px-6">
+      <div className="max-w-6xl mx-auto">
+        {/* Search, Filter, Sort Controls */}
+        {!loading && members.length > 0 && (
+          <div className="mb-8 space-y-4">
+            {/* Search and Sort Row */}
+            <div className="flex flex-col sm:flex-row gap-4">
+              {/* Search */}
+              <div className="relative flex-1">
+                <svg
+                  xmlns="http://www.w3.org/2000/svg"
+                  width="18"
+                  height="18"
+                  viewBox="0 0 24 24"
+                  fill="none"
+                  stroke="currentColor"
+                  strokeWidth="2"
+                  strokeLinecap="round"
+                  strokeLinejoin="round"
+                  className="absolute left-4 top-1/2 -translate-y-1/2 text-neutral-500"
+                  aria-hidden="true"
+                >
+                  <circle cx="11" cy="11" r="8" />
+                  <path d="m21 21-4.3-4.3" />
+                </svg>
+                <input
+                  type="text"
+                  value={searchQuery}
+                  onChange={(e) => setSearchQuery(e.target.value)}
+                  placeholder="Search by name, location, job, bio..."
+                  className="w-full pl-11 pr-4 py-3 bg-neutral-900 border border-neutral-800 rounded-lg text-white placeholder-neutral-500 focus:outline-none focus:ring-2 focus:ring-emerald-400 focus:border-transparent"
+                />
+              </div>
+
+              {/* Filter Toggle Button */}
+              <button
+                onClick={() => setShowFilters(!showFilters)}
+                className={`px-4 py-3 border rounded-lg font-medium transition-colors flex items-center gap-2 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-emerald-400 ${
+                  showFilters || activeFilterCount > 0
+                    ? "bg-emerald-500/10 border-emerald-500/50 text-emerald-400"
+                    : "bg-neutral-900 border-neutral-800 text-neutral-300 hover:border-neutral-700"
+                }`}
+              >
+                <svg
+                  xmlns="http://www.w3.org/2000/svg"
+                  width="18"
+                  height="18"
+                  viewBox="0 0 24 24"
+                  fill="none"
+                  stroke="currentColor"
+                  strokeWidth="2"
+                  strokeLinecap="round"
+                  strokeLinejoin="round"
+                  aria-hidden="true"
+                >
+                  <polygon points="22 3 2 3 10 12.46 10 19 14 21 14 12.46 22 3" />
+                </svg>
+                Filters
+                {activeFilterCount > 0 && (
+                  <span className="bg-emerald-500 text-white text-xs px-1.5 py-0.5 rounded-full">
+                    {activeFilterCount}
+                  </span>
+                )}
+              </button>
+
+              {/* Sort Dropdown */}
+              <select
+                value={sortBy}
+                onChange={(e) => setSortBy(e.target.value as SortOption)}
+                className="px-4 py-3 bg-neutral-900 border border-neutral-800 rounded-lg text-neutral-300 focus:outline-none focus:ring-2 focus:ring-emerald-400 focus:border-transparent cursor-pointer"
+              >
+                <option value="newest">Newest Members</option>
+                <option value="oldest">Oldest Members</option>
+                <option value="mostTalks">Most Talks</option>
+                <option value="mostEvents">Most Events</option>
+                <option value="mostPRs">Most Pull Requests</option>
+                <option value="name">Name (A-Z)</option>
+              </select>
+            </div>
+
+            {/* Filter Checkboxes */}
+            {showFilters && (
+              <div className="p-4 bg-neutral-900 border border-neutral-800 rounded-lg space-y-4">
+                {/* Member Type Filter */}
+                <div>
+                  <div className="flex items-center justify-between mb-3">
+                    <span className="text-sm font-medium text-neutral-300">
+                      Member type
+                    </span>
+                    {activeFilterCount > 0 && (
+                      <button
+                        onClick={handleClearFilters}
+                        className="text-xs text-neutral-400 hover:text-white transition-colors"
+                      >
+                        Clear all
+                      </button>
+                    )}
+                  </div>
+                  <div className="flex flex-wrap gap-2">
+                    <button
+                      onClick={() => setFilters((f) => ({ ...f, memberType: "all" }))}
+                      className={`inline-flex items-center gap-2 px-4 py-2.5 rounded-lg transition-colors border min-h-[44px] ${
+                        filters.memberType === "all"
+                          ? "bg-emerald-500/10 border-emerald-500/50 text-emerald-400"
+                          : "bg-neutral-800 border-neutral-700 text-neutral-400 hover:border-neutral-600"
+                      }`}
+                    >
+                      <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" aria-hidden="true">
+                        <path d="M16 21v-2a4 4 0 0 0-4-4H6a4 4 0 0 0-4 4v2" />
+                        <circle cx="9" cy="7" r="4" />
+                        <path d="M22 21v-2a4 4 0 0 0-3-3.87" />
+                        <path d="M16 3.13a4 4 0 0 1 0 7.75" />
+                      </svg>
+                      <span className="text-sm">All</span>
+                    </button>
+                    <button
+                      onClick={() => setFilters((f) => ({ ...f, memberType: "human" }))}
+                      className={`inline-flex items-center gap-2 px-4 py-2.5 rounded-lg transition-colors border min-h-[44px] ${
+                        filters.memberType === "human"
+                          ? "bg-emerald-500/10 border-emerald-500/50 text-emerald-400"
+                          : "bg-neutral-800 border-neutral-700 text-neutral-400 hover:border-neutral-600"
+                      }`}
+                    >
+                      <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" aria-hidden="true">
+                        <path d="M20 21v-2a4 4 0 0 0-4-4H8a4 4 0 0 0-4 4v2" />
+                        <circle cx="12" cy="7" r="4" />
+                      </svg>
+                      <span className="text-sm">Humans</span>
+                    </button>
+                    <button
+                      onClick={() => setFilters((f) => ({ ...f, memberType: "agent" }))}
+                      className={`inline-flex items-center gap-2 px-4 py-2.5 rounded-lg transition-colors border min-h-[44px] ${
+                        filters.memberType === "agent"
+                          ? "bg-purple-500/10 border-purple-500/50 text-purple-400"
+                          : "bg-neutral-800 border-neutral-700 text-neutral-400 hover:border-neutral-600"
+                      }`}
+                    >
+                      <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" aria-hidden="true">
+                        <rect x="3" y="11" width="18" height="10" rx="2" />
+                        <circle cx="12" cy="5" r="2" />
+                        <path d="M12 7v4" />
+                        <line x1="8" y1="16" x2="8" y2="16" />
+                        <line x1="16" y1="16" x2="16" y2="16" />
+                      </svg>
+                      <span className="text-sm">Agents</span>
+                    </button>
+                  </div>
+                </div>
+
+                {/* Connected Accounts Filter */}
+                <div>
+                  <span className="text-sm font-medium text-neutral-300 block mb-3">
+                    Filter by connected accounts
+                  </span>
+                  <div className="flex flex-wrap gap-3">
+                    <FilterCheckbox
+                      checked={filters.hasDiscord}
+                      onChange={(checked) =>
+                        setFilters((f) => ({ ...f, hasDiscord: checked }))
+                      }
+                      label="Discord"
+                      icon={
+                        <svg width="14" height="14" viewBox="0 0 24 24" fill="currentColor" aria-hidden="true">
+                          <path d="M20.317 4.37a19.791 19.791 0 0 0-4.885-1.515.074.074 0 0 0-.079.037c-.21.375-.444.864-.608 1.25a18.27 18.27 0 0 0-5.487 0 12.64 12.64 0 0 0-.617-1.25.077.077 0 0 0-.079-.037A19.736 19.736 0 0 0 3.677 4.37a.07.07 0 0 0-.032.027C.533 9.046-.32 13.58.099 18.057a.082.082 0 0 0 .031.057 19.9 19.9 0 0 0 5.993 3.03.078.078 0 0 0 .084-.028 14.09 14.09 0 0 0 1.226-1.994.076.076 0 0 0-.041-.106 13.107 13.107 0 0 1-1.872-.892.077.077 0 0 1-.008-.128 10.2 10.2 0 0 0 .372-.292.074.074 0 0 1 .077-.01c3.928 1.793 8.18 1.793 12.062 0a.074.074 0 0 1 .078.01c.12.098.246.198.373.292a.077.077 0 0 1-.006.127 12.299 12.299 0 0 1-1.873.892.077.077 0 0 0-.041.107c.36.698.772 1.362 1.225 1.993a.076.076 0 0 0 .084.028 19.839 19.839 0 0 0 6.002-3.03.077.077 0 0 0 .032-.054c.5-5.177-.838-9.674-3.549-13.66a.061.061 0 0 0-.031-.03zM8.02 15.33c-1.183 0-2.157-1.085-2.157-2.419 0-1.333.956-2.419 2.157-2.419 1.21 0 2.176 1.096 2.157 2.42 0 1.333-.956 2.418-2.157 2.418zm7.975 0c-1.183 0-2.157-1.085-2.157-2.419 0-1.333.955-2.419 2.157-2.419 1.21 0 2.176 1.096 2.157 2.42 0 1.333-.946 2.418-2.157 2.418z"/>
+                        </svg>
+                      }
+                    />
+                    <FilterCheckbox
+                      checked={filters.hasLinkedIn}
+                      onChange={(checked) =>
+                        setFilters((f) => ({ ...f, hasLinkedIn: checked }))
+                      }
+                      label="LinkedIn"
+                      icon={
+                        <svg width="14" height="14" viewBox="0 0 24 24" fill="currentColor" aria-hidden="true">
+                          <path d="M20.447 20.452h-3.554v-5.569c0-1.328-.027-3.037-1.852-3.037-1.853 0-2.136 1.445-2.136 2.939v5.667H9.351V9h3.414v1.561h.046c.477-.9 1.637-1.85 3.37-1.85 3.601 0 4.267 2.37 4.267 5.455v6.286zM5.337 7.433c-1.144 0-2.063-.926-2.063-2.065 0-1.138.92-2.063 2.063-2.063 1.14 0 2.064.925 2.064 2.063 0 1.139-.925 2.065-2.064 2.065zm1.782 13.019H3.555V9h3.564v11.452zM22.225 0H1.771C.792 0 0 .774 0 1.729v20.542C0 23.227.792 24 1.771 24h20.451C23.2 24 24 23.227 24 22.271V1.729C24 .774 23.2 0 22.222 0h.003z"/>
+                        </svg>
+                      }
+                    />
+                    <FilterCheckbox
+                      checked={filters.hasTwitter}
+                      onChange={(checked) =>
+                        setFilters((f) => ({ ...f, hasTwitter: checked }))
+                      }
+                      label="X"
+                      icon={
+                        <svg width="14" height="14" viewBox="0 0 24 24" fill="currentColor" aria-hidden="true">
+                          <path d="M18.244 2.25h3.308l-7.227 8.26 8.502 11.24H16.17l-5.214-6.817L4.99 21.75H1.68l7.73-8.835L1.254 2.25H8.08l4.713 6.231zm-1.161 17.52h1.833L7.084 4.126H5.117z"/>
+                        </svg>
+                      }
+                    />
+                    <FilterCheckbox
+                      checked={filters.hasGithub}
+                      onChange={(checked) =>
+                        setFilters((f) => ({ ...f, hasGithub: checked }))
+                      }
+                      label="GitHub"
+                      icon={
+                        <svg width="14" height="14" viewBox="0 0 24 24" fill="currentColor" aria-hidden="true">
+                          <path d="M12 0c-6.626 0-12 5.373-12 12 0 5.302 3.438 9.8 8.207 11.387.599.111.793-.261.793-.577v-2.234c-3.338.726-4.033-1.416-4.033-1.416-.546-1.387-1.333-1.756-1.333-1.756-1.089-.745.083-.729.083-.729 1.205.084 1.839 1.237 1.839 1.237 1.07 1.834 2.807 1.304 3.492.997.107-.775.418-1.305.762-1.604-2.665-.305-5.467-1.334-5.467-5.931 0-1.311.469-2.381 1.236-3.221-.124-.303-.535-1.524.117-3.176 0 0 1.008-.322 3.301 1.23.957-.266 1.983-.399 3.003-.404 1.02.005 2.047.138 3.006.404 2.291-1.552 3.297-1.23 3.297-1.23.653 1.653.242 2.874.118 3.176.77.84 1.235 1.911 1.235 3.221 0 4.609-2.807 5.624-5.479 5.921.43.372.823 1.102.823 2.222v3.293c0 .319.192.694.801.576 4.765-1.589 8.199-6.086 8.199-11.386 0-6.627-5.373-12-12-12z"/>
+                        </svg>
+                      }
+                    />
+                    <FilterCheckbox
+                      checked={filters.hasSubstack}
+                      onChange={(checked) =>
+                        setFilters((f) => ({ ...f, hasSubstack: checked }))
+                      }
+                      label="Substack"
+                      icon={
+                        <svg width="14" height="14" viewBox="0 0 24 24" fill="currentColor" aria-hidden="true">
+                          <path d="M22.539 8.242H1.46V5.406h21.08v2.836zM1.46 10.812V24L12 18.11 22.54 24V10.812H1.46zM22.54 0H1.46v2.836h21.08V0z"/>
+                        </svg>
+                      }
+                    />
+                    <FilterCheckbox
+                      checked={filters.hasWebsite}
+                      onChange={(checked) =>
+                        setFilters((f) => ({ ...f, hasWebsite: checked }))
+                      }
+                      label="Website"
+                      icon={
+                        <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
+                          <circle cx="12" cy="12" r="10" />
+                          <path d="M12 2a14.5 14.5 0 0 0 0 20 14.5 14.5 0 0 0 0-20" />
+                          <path d="M2 12h20" />
+                        </svg>
+                      }
+                    />
+                  </div>
+                </div>
+              </div>
+            )}
+
+            {/* Results count */}
+            <div className="text-sm text-neutral-500">
+              {filteredAndSortedMembers.length === members.length
+                ? `${members.length} member${members.length !== 1 ? "s" : ""}`
+                : `${filteredAndSortedMembers.length} of ${members.length} members`}
+            </div>
+          </div>
+        )}
+
+        {loading ? (
+          <div className="flex justify-center py-12">
+            <div className="animate-spin rounded-full h-8 w-8 border-t-2 border-b-2 border-white"></div>
+          </div>
+        ) : members.length === 0 ? (
+          <div className="text-center py-12">
+            <p className="text-neutral-400 text-lg mb-4">
+              No public profiles yet.
+            </p>
+            <p className="text-neutral-500">
+              Be the first to{" "}
+              <Link
+                href="/profile"
+                className="text-emerald-400 hover:text-emerald-300 underline"
+              >
+                make your profile public
+              </Link>
+              !
+            </p>
+          </div>
+        ) : filteredAndSortedMembers.length === 0 ? (
+          <div className="text-center py-12">
+            <p className="text-neutral-400 text-lg mb-4">
+              No members match your search.
+            </p>
+            <button
+              onClick={handleClearFilters}
+              className="text-emerald-400 hover:text-emerald-300 underline"
+            >
+              Clear filters
+            </button>
+          </div>
+        ) : (
+          <div className="grid sm:grid-cols-2 lg:grid-cols-3 gap-6">
+            {filteredAndSortedMembers.map((member) => (
+              <MemberCard key={member.uid} member={member} />
+            ))}
+          </div>
+        )}
+      </div>
+    </section>
+  );
+}
