+++ lib/utils.ts
@@ -1,0 +1,39 @@
+import { Timestamp } from "firebase/firestore";
+
+/**
+ * Get initials from a name string.
+ * Returns first letter of first and last name, or single letter if only one word.
+ */
+export function getInitials(name: string | null | undefined): string {
+  if (name) {
+    const parts = name.trim().split(/\s+/);
+    if (parts.length >= 2) {
+      return (parts[0][0] + parts[parts.length - 1][0]).toUpperCase();
+    }
+    return name[0].toUpperCase();
+  }
+  return "?";
+}
+
+/**
+ * Format a Firestore timestamp as a relative date string.
+ * Returns "just now", "Xm ago", "Xh ago", "Xd ago", or formatted date.
+ */
+export function formatRelativeDate(timestamp: Timestamp): string {
+  if (!timestamp?.toDate) return "";
+  const date = timestamp.toDate();
+  const now = new Date();
+  const diffMs = now.getTime() - date.getTime();
+  const diffMins = Math.floor(diffMs / 60000);
+  const diffHours = Math.floor(diffMs / 3600000);
+  const diffDays = Math.floor(diffMs / 86400000);
+
+  if (diffMins < 1) return "just now";
+  if (diffMins < 60) return `${diffMins}m ago`;
+  if (diffHours < 24) return `${diffHours}h ago`;
+  if (diffDays < 7) return `${diffDays}d ago`;
+  return date.toLocaleDateString("en-US", {
+    month: "short",
+    day: "numeric",
+  });
+}
